requirements.txt
# langchain==0.0.284
# python-dotenv==1.0.0
# streamlit==1.22.0
# unstructured==0.9.2
# tiktoken==0.4.0
# faiss-cpu==1.7.4
# libmagic==1.0
# python-magic==0.4.27
# python-magic-bin==0.4.14
# OpenAI == 0.28.0

## Newer versions 
# langchain==0.1.0
# python-dotenv==1.0.0
# streamlit==1.28.0
# unstructured==0.11.0
# tiktoken==0.5.0
# faiss-cpu==1.7.4
# openai==1.0.0

# langchain==0.1.0
# langchain-openai==0.0.5
# langchain-community==0.0.13
# python-dotenv==1.0.0
# streamlit==1.28.0
# unstructured==0.11.0
# chromadb==0.4.22
# openai==1.0.0

langchain
langchain-openai
langchain-community
python-dotenv
streamlit
unstructured
chromadb
openai